profile route saves new user to mongodb
