+++ src/server/index.js
@@ -9,7 +9,6 @@ const app = express();
 require("./routes/auth")(app);
 require("./routes/match")(app);
 require("./routes/profile")(app);
-require('./routes/profile_test')(app);
 
 app.get("*", (req, res) => {
   res.send("Home Page");
